==> revo/notebooks/4_task2.ipynb <==
import pandas as pd
import numpy as np

from sklearn.impute import KNNImputer
from sklearn.model_selection import train_test_split

from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import accuracy_score, classification_report
import xgboost as xgb
import lightgbm as lgb

# Import cleaned data
df=pd.read_csv('output/0_clean_arff/df_task2.csv')

df=df.drop(['description_sector','Quarter'],axis=1)
df=pd.get_dummies(df)
df = df.astype(float)

# Change the code_sector to work with certain ml functions below
df['code_sector'] = (df['code_sector'] - 1).astype(int)

X_cols=df[['code_sector']].columns
Y_cols=df.drop(['code_sector'],axis=1).columns

# Split dataset into train-test
X = df.drop(columns=['code_sector'])
y = df['code_sector']

# Split data into train and test sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Apply outlier and missings inputation to train and test separately to avoid leakage

def remove_outliers_iqr(df,IQR_multiplier):
    """
    Remove outliers from each column of a DataFrame using the Interquartile Range (IQR) method.
    Turns the outliers with None
    """
    df_no_outliers = df.copy()  # Create a copy of the DataFrame to avoid modifying the original
    
    # Select columns with float dtype
    float_cols = df.select_dtypes(include=['float']).columns

    # Iterate over each float column
    for column in df[float_cols].columns:
        # Calculate the first quartile (Q1) and third quartile (Q3) for the column
        Q1 = df[column].quantile(0.25)
        Q3 = df[column].quantile(0.75)
        
        # Calculate the IQR for the column
        IQR = Q3 - Q1
        
        # Define the lower and upper bounds for outliers for the column
        lower_bound = Q1 - IQR_multiplier * IQR
        upper_bound = Q3 + IQR_multiplier * IQR
        
        # Replace outliers with None for the column
        df_no_outliers[column] = df[column].mask((df[column] < lower_bound) | (df[column] > upper_bound), other=None)
    
    return df_no_outliers

X_train = remove_outliers_iqr(X_train,4)
X_test = remove_outliers_iqr(X_test,4)

# Inpute missing data with interpolation as it is time series data
X_train = X_train.interpolate(method='linear')
X_test = X_test.interpolate(method='linear')

# Inpute the remaining missing values (due to missing from the start or from the end of df)
imputer = KNNImputer(n_neighbors=5)
X_train = imputer.fit_transform(X_train)
X_test = imputer.fit_transform(X_test)

# For more comprehensible output
labels_list=['Transportation and warehousing','Wholesale trade','Manufacturing','Retail trade','Energy','Construction']

# Initialize and train the Random Forest Classifier
clf = RandomForestClassifier(n_estimators=100, random_state=42)
clf.fit(X_train, y_train)

# Predict on the test set
y_pred = clf.predict(X_test)

# Calculate accuracy
accuracy = accuracy_score(y_test, y_pred)
print(f'Accuracy: {accuracy:.2f}')

# Classification report
print(classification_report(y_test, y_pred,target_names=labels_list))


# Initialize and train the XGBoost Classifier
clf = xgb.XGBClassifier(objective='multi:softmax', num_class=6, random_state=42)
clf.fit(X_train, y_train)

# Predict on the test set
y_pred = clf.predict(X_test)

# Calculate accuracy
accuracy = accuracy_score(y_test, y_pred)
print(f'Accuracy: {accuracy:.2f}')

# Classification report
print(classification_report(y_test, y_pred,target_names=labels_list))


# Initialize and train the LightGBM Classifier
clf = lgb.LGBMClassifier(num_class=6, random_state=42,force_col_wise=True)
clf.fit(X_train, y_train)

# Predict on the test set
y_pred = clf.predict(X_test)

# Calculate accuracy
accuracy = accuracy_score(y_test, y_pred)
print(f'Accuracy: {accuracy:.2f}')

# Classification report
print(classification_report(y_test, y_pred,target_names=labels_list))


# We dont export anything to output/ , we will export the best model with the production process.
